samples/function-calling/weather/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

google-cloud-aiplatform==1.87.0 # For main.py sample
google-genai==1.9.0 # For main_genaisdk.py sample